+++ super_editor/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [x.y.z] - ???
+ * FEATURE: For IME keyboard behavior - `option + backspace` deletes a word upstream, `option + delete` deletes a word downstream (#1161)
+
 ## [0.2.6] - May 28, 2023
  * FEATURE: `SuperReader` now launches URLs when tapping a link (#1151)
  * FIX: `SuperEditor` now correctly handles "\n" newlines reported by Android IME deltas (#1086)